Refactoring workspace search into a separate class
