+++ scripts/Muon/GUI/Common/plotting_widget/plotting_widget_presenter.py
@@ -10,13 +10,13 @@ from Muon.GUI.Common.home_tab.home_tab_presenter import HomeTabSubWidget
 from mantidqt.utils.observer_pattern import GenericObservable, GenericObserver, GenericObserverWithArgPassing
 from Muon.GUI.Common.utilities.run_string_utils import run_list_to_string
 from Muon.GUI.FrequencyDomainAnalysis.frequency_context import FREQUENCY_EXTENSIONS
+from Muon.GUI.Common.plotting_widget.workspace_finder import WorkspaceFinder
 
 COUNTS_PLOT_TYPE = 'Counts'
 ASYMMETRY_PLOT_TYPE = 'Asymmetry'
 FREQ_PLOT_TYPE = "Frequency "
-TIME_PLOT_TYPE = 'Time'
 workspace_index = 0
-x_limits = {FREQ_PLOT_TYPE: 50, TIME_PLOT_TYPE: 15}
+x_limits = {"Frequency": 50, "Time": 15}
 
 
 class PlotWidgetPresenter(HomeTabSubWidget):
@@ -30,27 +30,26 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         self._view = view
         self._model = model
         self.context = context
+        self.workspace_finder = WorkspaceFinder(self.context)
+
         self._view.on_rebin_options_changed(self.handle_use_raw_workspaces_changed)
         self._view.on_plot_type_changed(self.handle_plot_type_changed)
         self._view.on_tiled_by_type_changed(self.handle_tiled_by_changed_on_view)
         self._view.on_plot_tiled_changed(self.handle_plot_tiled_changed)
-        self.connect_xlim_changed_in_figure_view(self.handle_x_axis_limits_changed_in_figure_view)
-        self.connect_ylim_changed_in_figure_view(self.handle_y_axis_limits_changed_in_figure_view)
 
         self.input_workspace_observer = GenericObserver(self.handle_data_updated)
         self.added_group_or_pair_observer = GenericObserverWithArgPassing(
             self.handle_added_or_removed_group_or_pair_to_plot)
         self.fit_observer = GenericObserver(self.handle_fit_completed)
         self.removed_group_pair_observer = GenericObserver(self.handle_removed_group_or_pair_to_plot)
-
         self.rebin_options_set_observer = GenericObserver(self.handle_rebin_options_set)
         self.plot_guess_observer = GenericObserver(self.handle_plot_guess_changed)
         self.workspace_deleted_from_ads_observer = GenericObserverWithArgPassing(self.handle_workspace_deleted_from_ads)
         self.workspace_replaced_in_ads_observer = GenericObserverWithArgPassing(self.handle_workspace_replaced_in_ads)
-
         self.plot_type_changed_notifier = GenericObservable()
 
-        # plotting options
+        self.connect_xlim_changed_in_figure_view(self.handle_x_axis_limits_changed_in_figure_view)
+        self.connect_ylim_changed_in_figure_view(self.handle_y_axis_limits_changed_in_figure_view)
         self._view.get_plot_options().connect_errors_changed(self.handle_error_selection_changed)
         self._view.get_plot_options().connect_x_range_changed(self.handle_xlim_changed_in_options_view)
         self._view.get_plot_options().connect_y_range_changed(self.handle_ylim_changed_in_options_view)
@@ -86,6 +85,7 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         self._view.get_plot_options().set_plot_x_range([xmin, xmax])
         self._view.get_plot_options().set_plot_y_range([ymin, ymax])
         self._view.get_plot_options().clear_subplots()
+
         for i in range(self._model.number_of_axes):
             self._view.get_plot_options().add_subplot(str(i + 1))
 
@@ -98,13 +98,14 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         Handles the group, pair calculation finishing. Checks whether the list of workspaces has changed before doing
         anything as workspaces being modified in place is handled by the ADS handler observer.
         """
-        if self._model.plotted_workspaces == self.get_workspace_list_to_plot():
+        if self._model.plotted_workspaces == \
+                self.workspace_finder.get_workspace_list_to_plot(self._view.if_raw(), self._view.get_selected()):
             # If the workspace names have not changed the ADS observer should
             # handle any updates
             return
 
         if self._view.is_tiled_plot():
-            self.update_model_tile_plot_positions(self._view.get_tiled_by_type())
+            self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
 
         self.plot_all_selected_workspaces()
@@ -114,14 +115,13 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         Handles the workspace being deleted from ads
         """
         if self._view.is_tiled_plot():
-            self.update_model_tile_plot_positions(self._view.get_tiled_by_type())
+            self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
-
             self.plot_all_selected_workspaces()
 
         if workspace.name() in self._model.plotted_workspaces:
             self._model.workspace_deleted_from_ads(workspace, self._view.get_axes())
-            self._view.set_fig_titles(self.get_plot_title(self._view.is_tiled_plot(), self._view.get_tiled_by_type()))
+            self._view.set_fig_titles(self.get_plot_title())
             self._model.autoscale_axes(self._view.get_axes(), x_limits[self.get_domain()])
             self._view.force_redraw()
 
@@ -175,7 +175,7 @@ class PlotWidgetPresenter(HomeTabSubWidget):
                            'label': label}
             self._model.replot_workspace(workspace, ax, error_state, plot_kwargs)
 
-        self._model.autoscale_axes(self._view.get_axes(),x_limits[self.get_domain()])
+        self._model.autoscale_axes(self._view.get_axes(), x_limits[self.get_domain()])
         self._view.force_redraw()
 
     def handle_xlim_changed_in_options_view(self, xlim):
@@ -191,6 +191,7 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         for subplot in subplots:
             index = int(subplot) - 1
             self._model.set_axis_ylim(self._view.get_axes()[index], ylim[0], ylim[1])
+
         self._view.force_redraw()
 
     def handle_autoscale_requested_in_view(self):
@@ -200,8 +201,9 @@ class PlotWidgetPresenter(HomeTabSubWidget):
             xlims = self._view.get_plot_options().get_plot_x_range()
             indices = [ix - 1 for ix in list(map(int, subplots))]
             axes = [self._view.get_axes()[index] for index in indices]
+
             self._model.autoscale_axes(axes, xlims[1])
-            ymin,  ymax = self.get_y_lim()
+            ymin, ymax = self.get_y_lim(0)
             self._view.get_plot_options().set_plot_y_range([ymin, ymax])
             self._view.force_redraw()
 
@@ -225,27 +227,22 @@ class PlotWidgetPresenter(HomeTabSubWidget):
 
     def handle_subplot_changed_in_options_view(self):
         subplots = self._view.get_plot_options().get_selection()
-        if len(subplots) == 0:
+        if len(subplots) == 0 or len(subplots) > 1:
             return
         if subplots[0].isdigit():
-            if len(subplots) > 1:  # all is selected (get x and y lims from first subplot)
-                xmin, xmax = self.get_x_lim(0)
-                ymin, ymax = self.get_y_lim(0)
-                self._view.get_plot_options().set_plot_x_range([xmin, xmax])
-                self._view.get_plot_options().set_plot_y_range([ymin, ymax])
-            else:
-                index = int(subplots[0]) - 1
-                xmin, xmax = self.get_x_lim(index)
-                ymin, ymax = self.get_y_lim(index)
-                self._view.get_plot_options().set_plot_x_range([xmin, xmax])
-                self._view.get_plot_options().set_plot_y_range([ymin, ymax])
+            index = int(subplots[0]) - 1
+            xmin, xmax = self.get_x_lim(index)
+            ymin, ymax = self.get_y_lim(index)
+
+            self._view.get_plot_options().set_plot_x_range([xmin, xmax])
+            self._view.get_plot_options().set_plot_y_range([ymin, ymax])
 
     def handle_plot_tiled_changed(self, state):
         """
         Handles the plot tiled checkbox being changed in the view
         """
         if state == 2:  # tiled plot
-            self.update_model_tile_plot_positions(self._view.get_tiled_by_type())
+            self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
             self.plot_all_selected_workspaces()
         if state == 0:  # not tiled plot
@@ -276,7 +273,7 @@ class PlotWidgetPresenter(HomeTabSubWidget):
 
         # if tiled by group, we will need to recreate the tiles
         if self._view.is_tiled_plot() and self._view.get_tiled_by_type() == 'group':
-            self.update_model_tile_plot_positions(self._view.get_tiled_by_type())
+            self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
             self.plot_all_selected_workspaces()
         else:
@@ -289,17 +286,18 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         Handles a group or pair being removed in grouping widget analysis table
         """
         if self._view.is_tiled_plot() and self._view.get_tiled_by_type():
-            self.update_model_tile_plot_positions(self._view.get_tiled_by_type())
+            self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
             self.plot_all_selected_workspaces()
             return
 
-        workspace_list = self.get_workspace_list_to_plot()
+        workspace_list = self.workspace_finder.get_workspace_list_to_plot(self._view.if_raw(),
+                                                                          self._view.get_selected())
         for ws in self._model.plotted_workspaces:
             if ws not in workspace_list:
                 self._model.remove_workspace_from_plot(ws, self._view.get_axes())
 
-        self._model.autoscale_axes(self._view.get_axes(),x_limits[self.get_domain()])
+        self._model.autoscale_axes(self._view.get_axes(), x_limits[self.get_domain()])
         self._view.force_redraw()
 
     def handle_fit_completed(self):
@@ -346,14 +344,14 @@ class PlotWidgetPresenter(HomeTabSubWidget):
     def handle_tiled_by_changed_on_view(self, index):
         if self._view.is_tiled_plot():
             if index == 0:
-                self.update_model_tile_plot_positions('group')
+                self.update_model_tile_plot_positions()
             else:
-                self.update_model_tile_plot_positions('run')
+                self.update_model_tile_plot_positions()
             self.new_plot_figure(self._model.number_of_axes)
             self.plot_all_selected_workspaces()
 
     def handle_instrument_changed(self):
-        self.new_plot_figure(1)
+        self.new_plot_figure(num_axes=1)
 
     def handle_plot_guess_changed(self):
 
@@ -369,7 +367,6 @@ class PlotWidgetPresenter(HomeTabSubWidget):
                                                                'label': 'Fit Function Guess'})
         self._view.force_redraw()
 
-    # Connect to figure axis signals.
     # Every time a new figure is generated these signals will have
     # to be reconnected.
     def connect_xlim_changed_in_figure_view(self, slot):
@@ -388,10 +385,9 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         clearing any previous plots
         """
 
-        tiled = self._view.is_tiled_plot()
         errors = self._view.get_plot_options().get_errors()
-        # get the workspace list, formed from the selected groups / pairs
-        workspace_list = self.get_workspace_list_to_plot()
+        workspace_list = self.workspace_finder.get_workspace_list_to_plot(self._view.if_raw(),
+                                                                          self._view.get_selected())
 
         for ws in self._model.plotted_workspaces:
             if ws not in workspace_list:
@@ -406,8 +402,9 @@ class PlotWidgetPresenter(HomeTabSubWidget):
                                                                'label': label})
                 self._model.add_workspace_to_plotted_workspaces(workspace)
 
-        self._view.set_fig_titles(self.get_plot_title(tiled, self._view.get_tiled_by_type()))
-        self._model.autoscale_axes(self._view.get_axes(),x_limits[self.get_domain()])
+        # scale the axis and set title
+        self._view.set_fig_titles(self.get_plot_title())
+        self._model.autoscale_axes(self._view.get_axes(), x_limits[self.get_domain()])
         self.update_options_view_from_model()
         self._view.force_redraw()
 
@@ -416,17 +413,16 @@ class PlotWidgetPresenter(HomeTabSubWidget):
             for workspace in workspace_list
             if self.context.group_pair_context.get_equivalent_group_pair(workspace)}
 
-    def update_model_tile_plot_positions(self, tiled_type):
+    def update_model_tile_plot_positions(self):
         """
         Updates tile dictionary in the model, which maps workspaces to
         the axis which they are to be plotted on
-        :param tiled_type: A string which states whether the tiling is performed by run, or by group
         """
         self._model.tiled_plot_positions.clear()
         self._model.clear_plot_model(self._view.get_axes())
         self.context.fitting_context.clear()
 
-        if tiled_type == 'run':  # tile by run
+        if self._view.get_tiled_by_type() == 'run':  # tile by run
             flattened_run_list = [item for sublist in self.context.data_context.current_runs for item in sublist]
             instrument = self.context.data_context.instrument
             for i, run in enumerate(flattened_run_list):
@@ -445,79 +441,9 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         self.connect_xlim_changed_in_figure_view(self.handle_x_axis_limits_changed_in_figure_view)
         self.connect_ylim_changed_in_figure_view(self.handle_y_axis_limits_changed_in_figure_view)
 
-    def get_workspace_list_to_plot(self):
-        """
-         :return: a list of workspace names to plot
-         """
-        workspace_list = self.get_workspaces_to_plot(self._view.if_raw(), self._view.get_selected())
-        return workspace_list
-
-    def get_workspaces_to_plot(self, is_raw, plot_type):
-        """
-        :param is_raw: Whether to use raw or rebinned data
-        :param plot_type: plotting type, e.g Counts, Frequency Re
-        :return: a list of workspace names
-        """
-        currently_selected_groups = self.context.group_pair_context.selected_groups
-        currently_selected_pairs = self.context.group_pair_context.selected_pairs
-        workspace_list = []
-        if FREQ_PLOT_TYPE in plot_type:
-            for grouppair in currently_selected_groups + currently_selected_pairs:
-                workspace_list += self.get_freq_workspaces_to_plot(grouppair, plot_type)
-            return workspace_list
-        else:
-            for grouppair in currently_selected_groups + currently_selected_pairs:
-                workspace_list += self.get_time_workspaces_to_plot(grouppair, is_raw, plot_type)
-            return workspace_list
-
-    def get_freq_workspaces_to_plot(self, current_group_pair, plot_type):
-        """
-        :param current_group_pair: The group/pair currently selected
-        :param plot_type: Whether to plot counts or asymmetry
-        :return: a list of workspace names
-        """
-        try:
-            runs = ""
-            seperator = ""
-            for run in self.context.data_context.current_runs:
-                runs += seperator + str(run[0])
-                seperator = ", "
-            workspace_list = self.context.get_names_of_frequency_domain_workspaces_to_fit(
-                runs, current_group_pair, True, plot_type[len(FREQ_PLOT_TYPE):])
-
-            return workspace_list
-        except AttributeError:
-            return []
-
-    def get_time_workspaces_to_plot(self, current_group_pair, is_raw, plot_type):
-        """
-        :param current_group_pair: The group/pair currently selected
-        :param is_raw: Whether to use raw or rebinned data
-        :param plot_type: Whether to plot counts or asymmetry
-        :return: a list of workspace names
-        """
-        try:
-            if is_raw:
-                workspace_list = self.context.group_pair_context[current_group_pair].get_asymmetry_workspace_names(
-                    self.context.data_context.current_runs)
-            else:
-                workspace_list = self.context.group_pair_context[
-                    current_group_pair].get_asymmetry_workspace_names_rebinned(
-                    self.context.data_context.current_runs)
-
-            if plot_type == COUNTS_PLOT_TYPE:
-                workspace_list = [item.replace(ASYMMETRY_PLOT_TYPE, COUNTS_PLOT_TYPE)
-                                  for item in workspace_list if ASYMMETRY_PLOT_TYPE in item]
-
-            return workspace_list
-        except AttributeError:
-            return []
-
-    def get_plot_title(self, tiled, tiled_type):
+    def get_plot_title(self):
         """
         Generates a title for the plot based on current instrument group and run numbers
-        :param tiled: True or false as to whether this is a tiled plot
-        :param tiled_type: Whether we are tiling by group or by run
         :return: Plot titles
         """
         flattened_run_list = [
@@ -525,8 +451,8 @@ class PlotWidgetPresenter(HomeTabSubWidget):
         instrument = self.context.data_context.instrument
 
         plot_titles = []
-        if tiled:
-            if tiled_type == 'run':
+        if self._view.is_tiled_plot():
+            if self._view.get_tiled_by_type() == 'run':
                 for run in flattened_run_list:
                     plot_titles.append(instrument + str(run))
             else:
@@ -541,7 +467,7 @@ class PlotWidgetPresenter(HomeTabSubWidget):
 
     def get_workspace_legend_label(self, workspace_name):
         """
-        Generates a label for the workspace
+        Generates a label for the workspace which is used in the plot
         :return: workspace label
         """
         if self._view.is_tiled_plot():
@@ -563,19 +489,23 @@ class PlotWidgetPresenter(HomeTabSubWidget):
 
     def get_workspace_plot_axis(self, workspace_name):
         """
-        Retrieve the axis from the view, which the model will then use to plot the workspace
+        Returns the axis which the workspace will be plotted to
         :return: axis
         """
         if self._view.is_tiled_plot():
-            if self._view.get_tiled_by_type() == 'group':
-                tiled_key = self._get_group_or_pair_from_workspace_name(workspace_name)
-            else:
-                tiled_key = self.context.data_context.instrument + self._get_run_number_from_workspace(workspace_name)
+            tiled_key = self.get_tiled_key(workspace_name)
             position = self._model.tiled_plot_positions[tiled_key]
             ax = self._view.get_axes()[position]
         else:
             ax = self._view.get_axes()[0]
         return ax
+
+    def get_tiled_key(self, workspace_name):
+        if self._view.get_tiled_by_type() == 'group':
+            tiled_key = self._get_group_or_pair_from_workspace_name(workspace_name)
+        else:
+            tiled_key = self.context.data_context.instrument + self._get_run_number_from_workspace(workspace_name)
+        return tiled_key
 
     def get_domain(self):
         if FREQ_PLOT_TYPE in self._view.get_selected():
